Move method_config module from ext/client_channel to lib/transport.
