+++ BUILD
@@ -249,6 +249,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.h",
     "src/core/lib/transport/metadata.h",
     "src/core/lib/transport/metadata_batch.h",
+    "src/core/lib/transport/method_config.h",
     "src/core/lib/transport/static_metadata.h",
     "src/core/lib/transport/timeout_encoding.h",
     "src/core/lib/transport/transport.h",
@@ -306,7 +307,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.h",
     "src/core/ext/client_channel/lb_policy_factory.h",
     "src/core/ext/client_channel/lb_policy_registry.h",
-    "src/core/ext/client_channel/method_config.h",
     "src/core/ext/client_channel/parse_address.h",
     "src/core/ext/client_channel/resolver.h",
     "src/core/ext/client_channel/resolver_factory.h",
@@ -433,6 +433,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.c",
     "src/core/lib/transport/metadata.c",
     "src/core/lib/transport/metadata_batch.c",
+    "src/core/lib/transport/method_config.c",
     "src/core/lib/transport/static_metadata.c",
     "src/core/lib/transport/timeout_encoding.c",
     "src/core/lib/transport/transport.c",
@@ -499,7 +500,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.c",
     "src/core/ext/client_channel/lb_policy_factory.c",
     "src/core/ext/client_channel/lb_policy_registry.c",
-    "src/core/ext/client_channel/method_config.c",
     "src/core/ext/client_channel/parse_address.c",
     "src/core/ext/client_channel/resolver.c",
     "src/core/ext/client_channel/resolver_factory.c",
@@ -672,6 +672,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.h",
     "src/core/lib/transport/metadata.h",
     "src/core/lib/transport/metadata_batch.h",
+    "src/core/lib/transport/method_config.h",
     "src/core/lib/transport/static_metadata.h",
     "src/core/lib/transport/timeout_encoding.h",
     "src/core/lib/transport/transport.h",
@@ -706,7 +707,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.h",
     "src/core/ext/client_channel/lb_policy_factory.h",
     "src/core/ext/client_channel/lb_policy_registry.h",
-    "src/core/ext/client_channel/method_config.h",
     "src/core/ext/client_channel/parse_address.h",
     "src/core/ext/client_channel/resolver.h",
     "src/core/ext/client_channel/resolver_factory.h",
@@ -841,6 +841,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.c",
     "src/core/lib/transport/metadata.c",
     "src/core/lib/transport/metadata_batch.c",
+    "src/core/lib/transport/method_config.c",
     "src/core/lib/transport/static_metadata.c",
     "src/core/lib/transport/timeout_encoding.c",
     "src/core/lib/transport/transport.c",
@@ -882,7 +883,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.c",
     "src/core/ext/client_channel/lb_policy_factory.c",
     "src/core/ext/client_channel/lb_policy_registry.c",
-    "src/core/ext/client_channel/method_config.c",
     "src/core/ext/client_channel/parse_address.c",
     "src/core/ext/client_channel/resolver.c",
     "src/core/ext/client_channel/resolver_factory.c",
@@ -1050,6 +1050,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.h",
     "src/core/lib/transport/metadata.h",
     "src/core/lib/transport/metadata_batch.h",
+    "src/core/lib/transport/method_config.h",
     "src/core/lib/transport/static_metadata.h",
     "src/core/lib/transport/timeout_encoding.h",
     "src/core/lib/transport/transport.h",
@@ -1083,7 +1084,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.h",
     "src/core/ext/client_channel/lb_policy_factory.h",
     "src/core/ext/client_channel/lb_policy_registry.h",
-    "src/core/ext/client_channel/method_config.h",
     "src/core/ext/client_channel/parse_address.h",
     "src/core/ext/client_channel/resolver.h",
     "src/core/ext/client_channel/resolver_factory.h",
@@ -1211,6 +1211,7 @@ cc_library(
     "src/core/lib/transport/mdstr_hash_table.c",
     "src/core/lib/transport/metadata.c",
     "src/core/lib/transport/metadata_batch.c",
+    "src/core/lib/transport/method_config.c",
     "src/core/lib/transport/static_metadata.c",
     "src/core/lib/transport/timeout_encoding.c",
     "src/core/lib/transport/transport.c",
@@ -1252,7 +1253,6 @@ cc_library(
     "src/core/ext/client_channel/lb_policy.c",
     "src/core/ext/client_channel/lb_policy_factory.c",
     "src/core/ext/client_channel/lb_policy_registry.c",
-    "src/core/ext/client_channel/method_config.c",
     "src/core/ext/client_channel/parse_address.c",
     "src/core/ext/client_channel/resolver.c",
     "src/core/ext/client_channel/resolver_factory.c",
@@ -2136,6 +2136,7 @@ objc_library(
     "src/core/lib/transport/mdstr_hash_table.c",
     "src/core/lib/transport/metadata.c",
     "src/core/lib/transport/metadata_batch.c",
+    "src/core/lib/transport/method_config.c",
     "src/core/lib/transport/static_metadata.c",
     "src/core/lib/transport/timeout_encoding.c",
     "src/core/lib/transport/transport.c",
@@ -2202,7 +2203,6 @@ objc_library(
     "src/core/ext/client_channel/lb_policy.c",
     "src/core/ext/client_channel/lb_policy_factory.c",
     "src/core/ext/client_channel/lb_policy_registry.c",
-    "src/core/ext/client_channel/method_config.c",
     "src/core/ext/client_channel/parse_address.c",
     "src/core/ext/client_channel/resolver.c",
     "src/core/ext/client_channel/resolver_factory.c",
@@ -2354,6 +2354,7 @@ objc_library(
     "src/core/lib/transport/mdstr_hash_table.h",
     "src/core/lib/transport/metadata.h",
     "src/core/lib/transport/metadata_batch.h",
+    "src/core/lib/transport/method_config.h",
     "src/core/lib/transport/static_metadata.h",
     "src/core/lib/transport/timeout_encoding.h",
     "src/core/lib/transport/transport.h",
@@ -2411,7 +2412,6 @@ objc_library(
     "src/core/ext/client_channel/lb_policy.h",
     "src/core/ext/client_channel/lb_policy_factory.h",
     "src/core/ext/client_channel/lb_policy_registry.h",
-    "src/core/ext/client_channel/method_config.h",
     "src/core/ext/client_channel/parse_address.h",
     "src/core/ext/client_channel/resolver.h",
     "src/core/ext/client_channel/resolver_factory.h",
