+++ src/python/grpcio/grpc_core_dependencies.py
@@ -179,6 +179,7 @@ CORE_SOURCE_FILES = [
   'src/core/lib/transport/mdstr_hash_table.c',
   'src/core/lib/transport/metadata.c',
   'src/core/lib/transport/metadata_batch.c',
+  'src/core/lib/transport/method_config.c',
   'src/core/lib/transport/static_metadata.c',
   'src/core/lib/transport/timeout_encoding.c',
   'src/core/lib/transport/transport.c',
@@ -245,7 +246,6 @@ CORE_SOURCE_FILES = [
   'src/core/ext/client_channel/lb_policy.c',
   'src/core/ext/client_channel/lb_policy_factory.c',
   'src/core/ext/client_channel/lb_policy_registry.c',
-  'src/core/ext/client_channel/method_config.c',
   'src/core/ext/client_channel/parse_address.c',
   'src/core/ext/client_channel/resolver.c',
   'src/core/ext/client_channel/resolver_factory.c',
