+++ tools/doxygen/Doxyfile.core.internal
@@ -873,6 +873,7 @@ src/core/lib/transport/connectivity_state.h \
 src/core/lib/transport/mdstr_hash_table.h \
 src/core/lib/transport/metadata.h \
 src/core/lib/transport/metadata_batch.h \
+src/core/lib/transport/method_config.h \
 src/core/lib/transport/static_metadata.h \
 src/core/lib/transport/timeout_encoding.h \
 src/core/lib/transport/transport.h \
@@ -930,7 +931,6 @@ src/core/ext/client_channel/initial_connect_string.h \
 src/core/ext/client_channel/lb_policy.h \
 src/core/ext/client_channel/lb_policy_factory.h \
 src/core/ext/client_channel/lb_policy_registry.h \
-src/core/ext/client_channel/method_config.h \
 src/core/ext/client_channel/parse_address.h \
 src/core/ext/client_channel/resolver.h \
 src/core/ext/client_channel/resolver_factory.h \
@@ -1061,6 +1061,7 @@ src/core/lib/transport/connectivity_state.c \
 src/core/lib/transport/mdstr_hash_table.c \
 src/core/lib/transport/metadata.c \
 src/core/lib/transport/metadata_batch.c \
+src/core/lib/transport/method_config.c \
 src/core/lib/transport/static_metadata.c \
 src/core/lib/transport/timeout_encoding.c \
 src/core/lib/transport/transport.c \
@@ -1127,7 +1128,6 @@ src/core/ext/client_channel/initial_connect_string.c \
 src/core/ext/client_channel/lb_policy.c \
 src/core/ext/client_channel/lb_policy_factory.c \
 src/core/ext/client_channel/lb_policy_registry.c \
-src/core/ext/client_channel/method_config.c \
 src/core/ext/client_channel/parse_address.c \
 src/core/ext/client_channel/resolver.c \
 src/core/ext/client_channel/resolver_factory.c \
